remove manual external name configurations with AutoExternalNameConfiguration()
merge service-group-detection and kind-name-generation into groups.go

add ProblematicResources() at provider.go

drop config/auto_discovery.go post moving its logic
