+++ config/groups.go
@@ -17,8 +17,18 @@
 package config
 
 import (
+	"strings"
+
 	"github.com/crossplane/upjet/pkg/config"
+	"golang.org/x/text/cases"
+	"golang.org/x/text/language"
 )
+
+const (
+	maxKubernetesNameLength    = 63
+	maxGeneratedKindNameLength = maxKubernetesNameLength - len("List")
+)
+
 
 // GroupKindCalculator returns the correct service group and kind for a resource
 type GroupKindCalculator func(resource string) (group string, kind string)
@@ -374,7 +384,6 @@ var GroupMap = map[string]GroupKindCalculator{
 	"oci_network_firewall_network_firewall": func(name string) (string, string) {
 		return "networkfirewall", "NetworkFirewall"
 	},
-	
 
 	// Load Balancer Service
 	"oci_load_balancer_load_balancer": func(name string) (string, string) {
@@ -431,6 +440,523 @@ var GroupMap = map[string]GroupKindCalculator{
 	"oci_psql_backup":        func(name string) (string, string) { return "psql", "PsqlBackup" },
 	"oci_psql_configuration": func(name string) (string, string) { return "psql", "PsqlConfiguration" },
 	"oci_psql_db_system":     func(name string) (string, string) { return "psql", "PsqlDbSystem" },
+}
+
+// ServiceGroupDetector automatically determines the service group for a resource
+// based on its name pattern. This is used when GroupMap doesn't have an explicit entry.
+func ServiceGroupDetector(resourceName string) (group string, kind string) {
+	// Extract the service prefix (e.g., "oci_database_*" -> "database")
+	parts := strings.Split(resourceName, "_")
+
+	// Fallback case: Resources with len(parts)<2 are grouped into core service
+	if len(parts) < 2 {
+		group = "core"
+		return group, generateKindName(resourceName, group)
+	}
+
+	servicePrefix := parts[1]
+
+	// Special handling for core resources that should be split
+	if servicePrefix == "core" {
+		group = detectCoreServiceGroup(resourceName)
+		return group, generateKindName(resourceName, group)
+	}
+
+	// Multi-word / colliding prefixes use parts[2] to disambiguate, defaulting to
+	// the first known service for that prefix. Single-word services rely on the
+	// default fallback (group = servicePrefix) and are not listed here.
+	switch servicePrefix {
+	case "ai":
+		// Ai Data Platform, Ai Document, Ai Language, Ai Vision
+		if len(parts) > 2 {
+			switch parts[2] {
+			case "data", "dataplatform", "data_platform":
+				group = "aidataplatform"
+			case "document":
+				group = "aidocument"
+			case "language":
+				group = "ailanguage"
+			case "vision":
+				group = "aivision"
+			default:
+				// default to first known AI service
+				group = "aidataplatform"
+			}
+		} else {
+			group = "aidataplatform"
+		}
+
+	case "announcements":
+		// "Announcements Service": "announcements_service" -> "announcementsservice"
+		group = "announcementsservice"
+
+	case "api":
+		// "API Platform": "api_platform" -> "apiplatform"
+		group = "apiplatform"
+
+	case "apm":
+		// Apm, Apm Config, Apm Synthetics, Apm Traces
+		if len(parts) > 2 {
+			switch parts[2] {
+			case "config":
+				group = "apmconfig"
+			case "synthetics":
+				group = "apmsynthetics"
+			case "traces":
+				group = "apmtraces"
+			default:
+				group = "apm"
+			}
+		} else {
+			group = "apm"
+		}
+
+	case "capacity":
+		// "Capacity Management": "capacity_management" -> "capacitymanagement"
+		group = "capacitymanagement"
+
+	case "certificates":
+		// "Certificates Management": "certificates_management" -> "certificatesmanagement"
+		group = "certificatesmanagement"
+
+	case "cloud":
+		// Cloud Bridge, Cloud Guard, Cloud Migrations
+		if len(parts) > 2 {
+			switch parts[2] {
+			case "bridge":
+				group = "cloudbridge"
+			case "guard":
+				group = "cloudguard"
+			case "migrations":
+				group = "cloudmigrations"
+			default:
+				group = "cloudbridge" // default to first service
+			}
+		} else {
+			group = "cloudbridge"
+		}
+
+	case "cluster":
+		// "Cluster Placement Groups": "cluster_placement_groups" -> "clusterplacementgroups"
+		group = "clusterplacementgroups"
+
+	case "compute":
+		// "Compute Cloud At Customer": "compute_cloud_at_customer" -> "computecloudatcustomer"
+		group = "computecloudatcustomer"
+
+	case "container":
+		// Container Engine, Container Instances
+		if len(parts) > 2 {
+			switch parts[2] {
+			case "engine":
+				group = "containerengine"
+			case "instances":
+				group = "containerinstances"
+			default:
+				group = "containerengine"
+			}
+		} else {
+			group = "containerengine"
+		}
+
+	case "data":
+		// Data Labeling Service, Data Safe
+		if len(parts) > 2 {
+			switch parts[2] {
+			case "labeling":
+				group = "datalabelingservice"
+			case "safe":
+				group = "datasafe"
+			default:
+				group = "datasafe"
+			}
+		} else {
+			group = "datasafe"
+		}
+
+	case "delegate":
+		// Delegate Access Control: "delegate_access_control" -> "delegateaccesscontrol"
+		group = "delegateaccesscontrol"
+
+	case "demand":
+		// Demand Signal: "demand_signal" -> "demandsignal"
+		group = "demandsignal"
+
+	case "disaster":
+		// Disaster Recovery: "disaster_recovery" -> "disasterrecovery"
+		group = "disasterrecovery"
+
+	case "file":
+		// "File Storage": "file_storage" -> "filestorage"
+		group = "filestorage"
+
+	case "fleet":
+		// Fleet Apps Management, Fleet Software Update
+		if len(parts) > 2 {
+			switch parts[2] {
+			case "apps":
+				group = "fleetappsmanagement"
+			case "software":
+				group = "fleetsoftwareupdate"
+			default:
+				group = "fleetappsmanagement"
+			}
+		} else {
+			group = "fleetappsmanagement"
+		}
+
+	case "lustre":
+		// "Lustre File Storage": "lustre_file_storage" -> "lustrefilestorage"
+		group = "lustrefilestorage"
+
+	case "managed":
+		// "Managed Kafka": "managed_kafka" -> "managedkafka"
+		group = "managedkafka"
+
+	case "media":
+		// "Media Services": "media_services" -> "mediaservices"
+		group = "mediaservices"
+
+	case "metering":
+		// "Metering Computation": "metering_computation" -> "meteringcomputation"
+		group = "meteringcomputation"
+
+	case "generic":
+		// "Generic Artifacts Content": "generic_artifacts_content" -> "genericartifactscontent"
+		group = "genericartifactscontent"
+
+	case "globally":
+		// "Globally Distributed Database": "globally_distributed_database" -> "globallydistributeddatabase"
+		group = "globallydistributeddatabase"
+
+	case "golden":
+		// "Golden Gate": "golden_gate" -> "goldengate"
+		group = "goldengate"
+
+	case "fusion":
+		// "Fusion Apps": "fusion_apps" -> "fusionapps"
+		group = "fusionapps"
+
+	case "generative":
+		// Generative AI, Generative Ai Agent
+		group = "generativeai"
+
+	case "health":
+		// "Health Checks": "health_checks" -> "healthchecks"
+		group = "healthchecks"
+
+	case "identity":
+		// Identity, Identity Data Plane, Identity Domains
+		if len(parts) > 2 {
+			switch parts[2] {
+			case "data":
+				group = "identitydataplane"
+			case "domains":
+				group = "identitydomains"
+			default:
+				group = "identity"
+			}
+		} else {
+			group = "identity"
+		}
+
+	case "jms":
+		// Jms, Jms Java Downloads, Jms Utils
+		if len(parts) > 2 {
+			switch parts[2] {
+			case "java":
+				group = "jmsjavadownloads"
+			case "utils":
+				group = "jmsutils"
+			default:
+				group = "jms"
+			}
+		} else {
+			group = "jms"
+		}
+
+	case "license":
+		// "License Manager": "license_manager" -> "licensemanager"
+		group = "licensemanager"
+
+	case "load":
+		// "Load Balancer": "load_balancer" -> "loadbalancer"
+		group = "loadbalancer"
+
+	case "log":
+		// "Log Analytics": "log_analytics" -> "loganalytics"
+		group = "loganalytics"
+
+	case "management":
+		// Management Agent, Management Dashboard
+		if len(parts) > 2 {
+			switch parts[2] {
+			case "agent":
+				group = "managementagent"
+			case "dashboard":
+				group = "managementdashboard"
+			default:
+				group = "managementagent"
+			}
+		} else {
+			group = "managementagent"
+		}
+
+	case "network":
+		// Network Firewall, Network Load Balancer; default networking
+		if len(parts) > 2 {
+			switch parts[2] {
+			case "firewall":
+				group = "networkfirewall"
+			case "load":
+				group = "networkloadbalancer"
+			default:
+				group = "networking"
+			}
+		} else {
+			group = "networking"
+		}
+
+	case "osp":
+		// "Osp Gateway": "osp_gateway" -> "ospgateway"
+		group = "ospgateway"
+
+	case "operator":
+		// "Operator Access Control": "operator_access_control" -> "operatoraccesscontrol"
+		group = "operatoraccesscontrol"
+
+	case "os":
+		// "Os Management Hub": "os_management_hub" -> "osmanagementhub"
+		group = "osmanagementhub"
+
+	case "resource":
+		// Resource Analytics, Resource Manager, Resource Scheduler
+		if len(parts) > 2 {
+			switch parts[2] {
+			case "analytics":
+				group = "resourceanalytics"
+			case "manager":
+				group = "resourcemanager"
+			case "scheduler":
+				group = "resourcescheduler"
+			default:
+				group = "resourcemanager"
+			}
+		} else {
+			group = "resourcemanager"
+		}
+
+	case "security":
+		// "Security Attribute": "security_attribute" -> "securityattribute"
+		group = "securityattribute"
+
+	case "service":
+		// "Service Catalog": "service_catalog" -> "servicecatalog"
+		group = "servicecatalog"
+
+	case "stack":
+		// "Stack Monitoring": "stack_monitoring" -> "stackmonitoring"
+		group = "stackmonitoring"
+
+	case "usage":
+		// "Usage Proxy": "usage_proxy" -> "usageproxy"
+		group = "usageproxy"
+
+	case "vbs":
+		// "Vbs Inst": "vbs_inst" -> "vbsinst"
+		group = "vbsinst"
+
+	case "visual":
+		// "Visual Builder": "visual_builder" -> "visualbuilder"
+		group = "visualbuilder"
+
+	case "vn":
+		// "Vn Monitoring": "vn_monitoring" -> "vnmonitoring"
+		group = "vnmonitoring"
+
+	case "vulnerability":
+		// "Vulnerability Scanning": "vulnerability_scanning" -> "vulnerabilityscanning"
+		group = "vulnerabilityscanning"
+
+	default:
+		// Fallback: use the raw service prefix
+		group = servicePrefix
+	}
+
+	return group, generateKindName(resourceName, group)
+}
+
+// detectCoreServiceGroup intelligently splits oci_core_* resources into logical services
+func detectCoreServiceGroup(resourceName string) (group string) {
+	// Compute resources
+	if contains(resourceName, []string{"instance", "image", "dedicated_vm", "console", "shape", "app_catalog", "cluster_network", "compute_"}) {
+		return "compute"
+	}
+
+	// Networking resources
+	if contains(resourceName, []string{"vcn", "subnet", "vnic", "dhcp", "vlan", "gateway", "security", "route", "ip", "peering"}) {
+		return "networking"
+	}
+
+	// Block storage resources
+	if contains(resourceName, []string{"volume", "boot_volume"}) {
+		return "blockstorage"
+	}
+
+	// Network connectivity resources (DRG, IPSec, etc.)
+	if contains(resourceName, []string{"drg", "cross_connect", "virtual_circuit", "cpe", "ipsec"}) {
+		return "networkconnectivity"
+	}
+
+	// Monitoring resources
+	if contains(resourceName, []string{"capture_filter", "vtap"}) {
+		return "monitoring"
+	}
+
+	// Default to core for unmatched patterns
+	return "core"
+}
+
+// generateKindName converts a resource name to a Kind name by stripping the
+// OCI prefix and all tokens that make up the logical service group.
+//
+//   - resourceName: full Terraform resource name, e.g. "oci_os_management_update_schedule".
+//   - group: the service group returned by ServiceGroupDetector, e.g. "osmanagement".
+//
+// If the resource name does not have at least three underscore-separated parts,
+// the function returns the original resourceName unchanged.
+func generateKindName(resourceName, group string) string {
+	parts := strings.Split(resourceName, "_")
+	if len(parts) < 3 {
+		// Not in expected oci_<service>_<resource> form; return as-is.
+		return resourceName
+	}
+
+	tokens := parts[1:] // drop the first segment (usually "oci")
+	if len(tokens) == 0 {
+		return resourceName
+	}
+
+	// Find the shortest prefix of tokens whose concatenation matches the group
+	// string (case-insensitive). That prefix represents the service name
+	// portion to drop from the Kind.
+	groupLower := strings.ToLower(group)
+	prefixEnd := 0
+	if groupLower != "" {
+		for i := 1; i <= len(tokens); i++ {
+			candidate := strings.ToLower(strings.Join(tokens[:i], ""))
+			if candidate == groupLower {
+				prefixEnd = i
+				break
+			}
+		}
+	}
+
+	// If we couldn't match the group to a prefix, fall back to dropping
+	// only the first token (service prefix) as in the original behavior.
+	start := 1
+	if prefixEnd > 0 {
+		start = prefixEnd
+	}
+	if start >= len(tokens) {
+		// No resource tokens left; return original resourceName defensively.
+		return resourceName
+	}
+
+	kindTokens := make([]string, 0, len(tokens[start:]))
+	for _, token := range tokens[start:] {
+		if token == "" {
+			continue
+		}
+		kindTokens = append(kindTokens, token)
+	}
+	if len(kindTokens) == 0 {
+		return resourceName
+	}
+
+	titleCaser := cases.Title(language.Und)
+	var b strings.Builder
+	for _, t := range kindTokens {
+		b.WriteString(titleCaser.String(t))
+	}
+
+	if b.Len() == 0 {
+		return resourceName
+	}
+
+	kindName := b.String()
+	if len(kindName) > maxGeneratedKindNameLength {
+		return removeLongestDuplicateSubstring(kindTokens)
+	}
+	return kindName
+}
+// removeLongestDuplicateSubstring shortens long generated kinds by removing the
+// latest duplicated token window whose normalized form already appears earlier
+// in the same name. This keeps the logic generic for overlong kinds without
+// hardcoding resource-specific cases.
+func removeLongestDuplicateSubstring(kindTokens []string) string {
+	titleCaser := cases.Title(language.Und)
+	render := func(tokens []string) string {
+		var b strings.Builder
+		for _, token := range tokens {
+			b.WriteString(titleCaser.String(token))
+		}
+		return b.String()
+	}
+	tokens := append([]string(nil), kindTokens...)
+	kindName := render(tokens)
+	for len(kindName) > maxGeneratedKindNameLength {
+		bestStart := -1
+		bestEnd := -1
+		bestLength := 0
+		bestWindowSize := 0
+		for start := 1; start < len(tokens); start++ {
+			prefix := strings.ToLower(strings.Join(tokens[:start], ""))
+			if prefix == "" {
+				continue
+			}
+			var candidate strings.Builder
+			for end := start; end < len(tokens); end++ {
+				candidate.WriteString(strings.ToLower(tokens[end]))
+				normalized := candidate.String()
+				if normalized == "" || !strings.Contains(prefix, normalized) {
+					continue
+				}
+				if start > 0 && end < len(tokens)-1 {
+					left := strings.ToLower(tokens[start-1])
+					right := strings.ToLower(tokens[end+1])
+					if left != "" && right != "" && (strings.Contains(left, right) || strings.Contains(right, left)) {
+						continue
+					}
+				}
+				windowSize := end - start + 1
+				if len(normalized) > bestLength ||
+					(len(normalized) == bestLength && windowSize > bestWindowSize) ||
+					(len(normalized) == bestLength && windowSize == bestWindowSize && start > bestStart) {
+					bestStart = start
+					bestEnd = end
+					bestLength = len(normalized)
+					bestWindowSize = windowSize
+				}
+			}
+		}
+		if bestStart == -1 {
+			return kindName
+		}
+		tokens = append(tokens[:bestStart], tokens[bestEnd+1:]...)
+		kindName = render(tokens)
+	}
+	return kindName
+}
+
+// contains checks if the resource name contains any of the given patterns
+func contains(resourceName string, patterns []string) bool {
+	for _, pattern := range patterns {
+		if strings.Contains(resourceName, pattern) {
+			return true
+		}
+	}
+	return false
 }
 
 // GroupKindOverrides applies the GroupMap during provider configuration
